+++ docs/client/docs.js
@@ -223,7 +223,8 @@ var toc = [
         {instance: "context", name: "invalidate"}
       ],
       {name: "Meteor.deps.Context.current", id: "current"},
-      "Meteor.flush"
+      "Meteor.flush",
+      "Meteor.autorun"
     // ],
 
     // "Environment Variables", [